requirements.txt
nextcord
typing
function-cooldowns
aiohttp
asyncio
requests
pymysql
wikipedia-api
matplotlib
pillow
datetime
quickchart-io
numpy